refactor: kakaomap into commons
