+++ src/components/common/TheKakaoMap.vue
@@ -8,8 +8,8 @@ const props = defineProps({
   positions: {
     type: Array,
     default: () => [],
-  }
-})
+  },
+});
 
 const kakaoStore = useKakaoStore();
 let mapBox = ref(null);
@@ -19,7 +19,10 @@ let markers = [];
 
 const initializeMap = (kakao) => {
   const options = {
-    center: new kakao.maps.LatLng(kakaoStore.getCoordinate[0], kakaoStore.getCoordinate[1]), // 초기 중심 좌표
+    center: new kakao.maps.LatLng(
+      kakaoStore.getCoordinate[0],
+      kakaoStore.getCoordinate[1]
+    ), // 초기 중심 좌표
     level: 3, // 확대 레벨
   };
 
@@ -34,30 +37,31 @@ const initializeMap = (kakao) => {
 };
 
 /**
- * 
- * @param {관광지정보} positions 
- * @param {카카오객체} kakao 
- * 
+ *
+ * @param {관광지정보} positions
+ * @param {카카오객체} kakao
+ *
  * @return 카카오 맵에 마커 찍기
  */
 const addMarkers = (positions, kakao) => {
   clearMarkers();
-  const imageSrc = "https://t1.daumcdn.net/localimg/localimages/07/mapapidoc/markerStar.png";
+  const imageSrc =
+    "https://t1.daumcdn.net/localimg/localimages/07/mapapidoc/markerStar.png";
   const imageSize = new kakao.maps.Size(24, 35);
   const markerImage = new kakao.maps.MarkerImage(imageSrc, imageSize);
 
-  positions.forEach(m => {
+  positions.forEach((m) => {
     const marker = new kakao.maps.Marker({
       map: map,
       position: new kakao.maps.LatLng(m.latlng.Ma, m.latlng.La),
       title: m.title,
-      image: markerImage
-    })
+      image: markerImage,
+    });
     marker.setMap(map);
     markers.push(marker);
-  })
+  });
   map.setCenter(positions[0].latlng);
-}
+};
 
 /**
  * @retrun 기존 마커지우기
@@ -72,7 +76,6 @@ onMounted(async () => {
   try {
     const kakao = await loadKakaoMap();
     initializeMap(kakao); // 맵 초기화
-
   } catch (error) {
     console.error("카카오 맵 로드 중 오류 발생:", error);
   }
@@ -90,13 +93,15 @@ watch(
   () => kakaoStore.getCoordinate,
   (newCoordinate) => {
     if (map) {
-      const newLatLng = new kakao.maps.LatLng(newCoordinate[0], newCoordinate[1]);
+      const newLatLng = new kakao.maps.LatLng(
+        newCoordinate[0],
+        newCoordinate[1]
+      );
       map.setCenter(newLatLng);
       map.setLevel(1);
     }
   }
-)
-
+);
 </script>
 
 <template>
